making Items Deletable & using IDs
